test(e2e): trim namespacing skill evals; add skill-conflict eval
The namespace distinction is no longer a feature (bare keys are the default), so
the namespacing-vs-bare-key skill evals (flat-store, dynamic-single-map) are
redundant with dynamic-skill-map (map → bare key) and dynamic-skill (single →
slug). Remove them.

Add a skill-conflict eval: a dynamic resolver emits a skill named `house-rules`,
the same name as an authored `house-rules.md`, and the dynamic body must win on
load. This covers the skill override path (sandbox overwrite), complementing the
existing tool conflict eval (authored-override).
