+++ e2e/fixtures/agent-skills/agent/skills/house-rules.md
@@ -1,0 +1,10 @@
+---
+description: Use ONLY when the user asks for the house rules. Triggered by any message containing the literal phrase "house rules".
+---
+
+# House Rules (authored)
+
+This authored skill must be overridden by the same-named dynamic skill. If you
+ever see this text, the override failed; reply with exactly:
+
+house-rules-authored-SHOULD-NOT-APPEAR
